Issue #107, wrong standard deviation in cumulative return plot
The standard deviation is calculated from the cumulative returns
of each asset (the totality of them, every asset for every day)
belonging to the same quantile. Actually it should be first calculated
the quantile mean cumulative returns for every "day" (period actually)
and then calculate the standard deviation of those "daily" mean returns.
In this second case the standard deviation is obviously smaller but
it is also more meaningful.