Improved tests for perf.average_cumulative_return_by_quantile (#111)
* bug fixes in perf.average_cumulative_return_by_quantile

- Cannot apply demeaning on the daily returns and build the cumulative
returns afterward. Instead we have to build the cumulative returns
from the daily returns and then apply the demeaning (and finally
averaging them).

- Deamining in the Group Specific Breakdown is not performed

* Improved tests for perf.average_cumulative_return_by_quantile

Followed @twiecki suggestion:
"prefer tests that have small and simple inputs that are so
trivial that the correct answer is obvious"
The tests are now performed on 4 assets having a fixed return rate each
so that the results can be easily verified.