Added support for group neutral factor analysis
Alphalens analysis on a group neutral factor is currently limited. The main
shortcoming is the return analysis that this commit improved allow the
forward returns to be demeaned by group.

There were also 2 bugs:
- API documentation claimed that perf.mean_return_by_quantile was performing
the demeaning at group level, but it wasn't
- The same goes for perf.mean_information_coefficient

Also changed the API arguments names in a consistent way:
- the tears.* functions use 'long_short' and 'group_neutral' all around, as
those functions are intended to be the top level ones. 'by_group' is used
if the function support multiple plots output, one for each group
- the remaining API (mostry performance.*) use 'demeaned' and
 'group_adjust' as they used to be