Make cumulative returns computation more robust
Most of Alphalens functions are now factor frequency agnostic (the
factor can be computed daily, weekly, randomly) and they don't assume
forward returns periods to be multiple of the frequency at which a
factor is computed, but there are still some functions that rely on
those assumptions. This commit  addresses those limitations.

The main function which requires a fix is 'cumulative_returns', which
assumes returns to be computed at a specific frequency (daily, weekly,
etc) without  gaps (e.g. an event which can happend randomly doesn't
work well with current implementation) and the return periods must be
multiple of factor computation frequency.

In order to remove those limitation from 'cumulative_returns' the period
length for which the returns are computed is required and for this
reason is now included in the forward columns name, which now contains
the period lenght (e.g. 1 day, is now '1D', 3 days is now '3D', 2 hours
'2h' and so on).

Other small fixed are included in this commit which are related to
the removal of the above assumptions.

This commit also add tests for performance.cumulative_returns.