BUG: Figures are not closed causing memory leak
From matplotlib documentation:
The memory required for a Figure is not completely released until the
figure is explicitly closed with close(). Deleting all references to
the figure, and/or using the window manager to kill the window in
which the figure appears on the screen, is not enough, because
pyplot maintains internal references until close() is called.