Modify factor_2_data to allow for binning_by_group
The previous factor_2_data set raises a non-unique bin edges error
when executing utils.quantize_factor. This was not due to a code
error in the function. It was just a result of not having enough
data in each group when calling pd.qcut.

To create the dataset, the nan's were left in place from factor_1_data,
and the non-nan values were just shuffled around.