ENH: added rate of return option in 'create_event_study_tear_sheet'
Rate of return allows to compare return happening at different time
frequency though, if an event doesn't happen frequently, we cannot
enter new positions any single day, so the rate of return doesn't
work well in this scenario and the return is more sensible in this case.
For this reason it is now optional.